fix: move save changes and option buttons to the new header
